# src/Web.Api/Controllers/Workshops/Contracts/GetWorkshopForEditRequest.cs
namespace Web.Api.Controllers.Workshops.Contracts;

public record GetWorkshopForEditResponse(
    string Name,
    string Address,
    string RegistrationDate,
    string NationalId,
    string? PostalCode = null);