# src/Web.Api/Controllers/Workshops/Contracts/UpdateWorkshopRequest.cs
namespace Web.Api.Controllers.Workshops.Contracts;

public record UpdateWorkshopRequest(
    string Name,
    string Address,
    PersianDate RegistrationDate,
    string NationalId,
    string? PostalCode = null);