# src/Web.Ui/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "WageCore",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/workshops": {
      "post": {
        "tags": [
          "Workshop"
        ],
        "summary": "CreateWorkshop",
        "operationId": "CreateWorkshop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateWorkshopRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateWorkshopRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateWorkshopRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkshopCommandResponseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkshopCommandResponseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkshopCommandResponseResult"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Workshop"
        ],
        "summary": "GetUserWorkshops",
        "operationId": "GetUserWorkshops",
        "parameters": [
          {
            "name": "Pagination.PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Pagination.PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Pagination.Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Region",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [
                "Normal",
                "LessDeveloped"
              ],
              "enum": [
                0,
                1
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserWorkshopsResponsePagedResultResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserWorkshopsResponsePagedResultResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserWorkshopsResponsePagedResultResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workshops/{workshopId}": {
      "put": {
        "tags": [
          "Workshop"
        ],
        "summary": "UpdateWorkshop",
        "operationId": "UpdateWorkshop",
        "parameters": [
          {
            "name": "workshopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateWorkshopRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateWorkshopRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateWorkshopRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workshop"
        ],
        "summary": "DeleteWorkshop",
        "operationId": "DeleteWorkshop",
        "parameters": [
          {
            "name": "workshopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workshops/{workshopId}/edit": {
      "get": {
        "tags": [
          "Workshop"
        ],
        "summary": "GetWorkshopForEdit",
        "operationId": "GetWorkshopForEdit",
        "parameters": [
          {
            "name": "workshopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopForEditResponseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopForEditResponseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkshopForEditResponseResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BooleanResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "type": "boolean"
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CreateWorkshopCommandResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "workshopId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateWorkshopCommandResponseResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateWorkshopCommandResponse"
              }
            ]
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "CreateWorkshopRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Normal",
              "LessDeveloped"
            ],
            "enum": [
              0,
              1
            ]
          },
          "registrationDate": {
            "type": "string"
          },
          "nationalId": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "GetUserWorkshopsResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "nationalId": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Normal",
              "LessDeveloped"
            ],
            "enum": [
              0,
              1
            ]
          },
          "displayRegistrationDate": {
            "type": "string",
            "nullable": true
          },
          "employeesCount": {
            "type": "integer",
            "format": "int32"
          },
          "departmentsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "GetUserWorkshopsResponsePagedResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GetUserWorkshopsResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "readOnly": true,
            "format": "int32"
          }
        }
      },
      "GetUserWorkshopsResponsePagedResultResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/GetUserWorkshopsResponsePagedResult"
              }
            ]
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "GetWorkshopForEditResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Normal",
              "LessDeveloped"
            ],
            "enum": [
              0,
              1
            ]
          },
          "registrationDate": {
            "type": "string",
            "nullable": true
          },
          "nationalId": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GetWorkshopForEditResponseResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "badResultType": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "x-enumNames": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ],
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "response": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/GetWorkshopForEditResponse"
              }
            ]
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "UpdateWorkshopRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [
              "Normal",
              "LessDeveloped"
            ],
            "enum": [
              0,
              1
            ]
          },
          "registrationDate": {
            "type": "string"
          },
          "nationalId": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter token without Bearer",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}